Split validation tests to its own file, parametrize traits
